debug logic handle visibility
