+++ core/main.ts
@@ -30,7 +30,6 @@ export * from "./src/net/DelaySyncer"
 export * from "./src/net/NetProtocol"
 export * from "./src/ecs/systems/ui/ChatBubbleSystem"
 export * from "./src/ecs/systems/ui/ClickableSystem"
-export * from "./src/ecs/systems/ui/DebugSystem"
 export * from "./src/ecs/systems/ui/HealthBarSystem"
 export * from "./src/ecs/systems/ui/InputSystem"
 export * from "./src/ecs/systems/ui/NametagSystem"
